@@ -11,15 +11,16 @@ classifiers = [
     'Operating System :: OS Independent',
     'Programming Language :: Python',
     'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.8',
     'Programming Language :: Python :: 3.9',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
+    'Programming Language :: Python :: 3.12',
+    'Programming Language :: Python :: 3.13',
     'Programming Language :: Python :: 3 :: Only',
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 pydantic = "^2.0.0"
 fastapi = ">=0.100.0"
 typing-extensions = "^4.4.0"
@@ -29,12 +30,11 @@ tortoise-orm = { version = ">=0.20,<0.24", optional = true }
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.2,<9.0"
 pytest-cov = ">=4,<6"
-pytest-asyncio = ">=0.20.3,<0.25.0"
+pytest-asyncio = ">=0.20.3,<0.25.3"
 httpx = ">=0.23.3,<0.29.0"
 asgi-lifespan = "^2.0.0"
-ruff = ">=0.5.1,<0.8.6"
+ruff = ">=0.5.1,<0.9.5"
 mypy = ">=0.991,<1.15"
-black = ">=22.12,<25.0"
 pre-commit = ">=2.21,<4.0"
 uvicorn = ">=0.20,<0.34"
 dirty-equals = ">=0.6,<0.9"
@@ -51,39 +51,56 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
+asyncio_default_fixture_loop_scope = "session"
+
 addopts = """
 """
 
-[tool.black]
-line-length = 120
-target-version = ['py38']
-include = '\.pyi?$'
-exclude = '''
-/(
-    \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | _build
-  | buck-out
-  | build
-  | dist
-  # The following are specific to Black, you probably don't want those.
-  | blib2to3
-  | tests/data
-  | profiling
-)/
-'''
-
 [tool.ruff]
 line-length = 120
-target-version = "py38"
+target-version = "py39"
 
-[tool.lint.ruff]
-select = ["E", "F"]
-ignore = []
+[tool.ruff.lint]
+select = [
+    "ALL",
+]
+ignore = [
+     # single rule exclude
+    "B008",    # do not perform function calls in argument defaults
+    "S101",    # use of assert detected
+    "SIM114",  # allow have multiplie ifs with same body
+    "TRY003",  # allow long exc message
+    "RUF100",  # allow unused noqa
+    "RUF012",  # allow mutable class var
+    "COM812",  # will be handled by ruff format
+    "ISC001",  # will be handled by ruff format
+    "SLF001",  # allow access to private member
+    "N806",    # allow variable name to start from uppercase letter
+    "N805",    # ignore first argument name check
+    "N818",    # ignore exception name should ends with Error suffix
+    "FAST002", # allow to use non-annotated dependencies
+    "PLR0913",  # allow more than 5 arguments
+    "PLR2004", # allow to compare with "magic numbers"
+    "UP007",   # ignore X | Y annotation suggestion
+    "TC001",   # ignore moving things to if TYPE_CHECKING block
+    "TC002",   # ignore moving things to if TYPE_CHECKING block
+    "TC003",   # ignore moving things to if TYPE_CHECKING block
+    "A005",    # allow module name to shadow a Python standard-library module
+    "PLW2901", # ignore loop variable overwrite
+    "S602",    # ignore subprocess.popen with shell=True
+    "S307",    # ignore literal eval
+    "ERA001",  # ignore commented-out code
+
+    # full rule exclude
+    "D",    # ignore docstrings
+    "FA",  # ignore future annotations
+    "ANN", # ignore annotations
+    "FIX", # ignore FIXMEs
+    "TD",  # ignore all TODOs
+    "ARG", # ignore unused arguments
+    "FBT", # ignore flake8-boolean-trap
+    "EM",  # ignore error message formatting
+]
 exclude = [
     ".bzr",
     ".direnv",
@@ -113,7 +130,7 @@ dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 max-complexity = 10
 
 [tool.mypy]
-python_version = 3.8
+python_version = "3.9"
 strict = true
 follow_imports = "normal"
 ignore_missing_imports = true
