@@ -1,20 +1,28 @@
 from datetime import datetime
-from typing import List, Any
+from typing import Any
 from uuid import UUID, uuid4
 
 from faker import Faker
-from fastapi import FastAPI, Depends
+from fastapi import Depends, FastAPI
 from pydantic import BaseModel
-from sqlalchemy import text, select, func, Integer
+from sqlalchemy import Integer, func, select, text
 from sqlalchemy.dialects.postgresql import ARRAY
-from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, MappedAsDataclass
+from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+from sqlalchemy.orm import DeclarativeBase, Mapped, MappedAsDataclass, mapped_column
 
-from fastapi_filters import SortingValues, FilterSet, FilterField, create_filters_from_set, create_sorting
+from fastapi_filters import (
+    FilterField,
+    FilterSet,
+    SortingValues,
+    create_filters_from_set,
+    create_sorting,
+)
 from fastapi_filters.ext.sqlalchemy import apply_filters_and_sorting
 
 faker = Faker()
-engine = create_async_engine("postgresql+asyncpg://postgres:postgres@localhost:5432/postgres")
+engine = create_async_engine(
+    "postgresql+asyncpg://postgres:postgres@localhost:5432/postgres",
+)
 
 
 class Base(MappedAsDataclass, DeclarativeBase):
@@ -30,7 +38,7 @@ class User(Base):
     last_name: Mapped[str] = mapped_column()
     email: Mapped[str] = mapped_column()
     age: Mapped[int] = mapped_column()
-    marks: Mapped[List[int]] = mapped_column(ARRAY(Integer))
+    marks: Mapped[list[int]] = mapped_column(ARRAY(Integer))
     is_active: Mapped[bool] = mapped_column()
     created_at: Mapped[datetime] = mapped_column()
 
@@ -41,7 +49,7 @@ class UserOut(BaseModel):
     last_name: str
     email: str
     age: int
-    marks: List[int]
+    marks: list[int]
     is_active: bool
     created_at: datetime
 
@@ -61,25 +69,24 @@ async def on_startup() -> None:
 
         await conn.run_sync(Base.metadata.create_all)
 
-    async with AsyncSession(engine) as session:
-        async with session.begin():
-            session.add_all(
-                [
-                    User(
-                        id=uuid4(),
-                        first_name=faker.first_name(),
-                        last_name=faker.last_name(),
-                        email=faker.email(),
-                        marks=[
-                            faker.pyint(min_value=1, max_value=5) for _ in range(faker.pyint(min_value=1, max_value=10))
-                        ],
-                        age=faker.pyint(min_value=1, max_value=100),
-                        is_active=faker.pybool(),
-                        created_at=faker.date_time_this_year(),
-                    )
-                    for _ in range(100)
-                ],
-            )
+    async with AsyncSession(engine) as session, session.begin():
+        session.add_all(
+            [
+                User(
+                    id=uuid4(),
+                    first_name=faker.first_name(),
+                    last_name=faker.last_name(),
+                    email=faker.email(),
+                    marks=[
+                        faker.pyint(min_value=1, max_value=5) for _ in range(faker.pyint(min_value=1, max_value=10))
+                    ],
+                    age=faker.pyint(min_value=1, max_value=100),
+                    is_active=faker.pybool(),
+                    created_at=faker.date_time_this_year(),
+                )
+                for _ in range(100)
+            ],
+        )
 
 
 class UserFiltersSet(FilterSet):
@@ -87,12 +94,12 @@ class UserFiltersSet(FilterSet):
     last_name: FilterField[str]
     email: FilterField[str]
     age: FilterField[int]
-    marks: FilterField[List[int]]
+    marks: FilterField[list[int]]
     is_active: FilterField[bool]
     created_at: FilterField[datetime]
 
 
-@app.get("/users", response_model=List[UserOut])
+@app.get("/users", response_model=list[UserOut])
 async def get_users(
     filters: UserFiltersSet = Depends(create_filters_from_set(UserFiltersSet)),
     sorting: SortingValues = Depends(create_sorting("age", "created_at")),
