@@ -1,8 +1,9 @@
-from typing import TypeVar, Any, Union, Optional, Mapping
+from collections.abc import Mapping
+from typing import Any, Optional, TypeVar, Union, cast
 
 from tortoise.queryset import QuerySet
 
-from fastapi_filters import FilterValues, FilterSet, FilterOperator
+from fastapi_filters import FilterOperator, FilterSet, FilterValues
 from fastapi_filters.types import AbstractFilterOperator, SortingValues
 
 TStmt = TypeVar("TStmt", bound=QuerySet[Any])
@@ -43,7 +44,7 @@ def apply_filters(
 
         for op, val in field_filters.items():
             if (cond := DEFAULT_FILTERS.get(op)) is not None:
-                stmt = stmt.filter(**{f"{field}{cond}": val})  # type: ignore
+                stmt = cast(TStmt, stmt.filter(**{f"{field}{cond}": val}))
             else:
                 raise NotImplementedError(f"Operator {op} is not implemented")
 
@@ -63,7 +64,7 @@ def apply_sorting(
         ordering.append(f"{'-' if direction == 'desc' else ''}{field}")
 
     if ordering:
-        stmt = stmt.order_by(*ordering)  # type: ignore
+        stmt = stmt.order_by(*ordering)  # type: ignore[assignment]
 
     return stmt
 
@@ -76,13 +77,11 @@ def apply_filters_and_sorting(
     remapping: Optional[Mapping[str, str]] = None,
 ) -> TStmt:
     stmt = apply_filters(stmt, filters, remapping=remapping)
-    stmt = apply_sorting(stmt, sorting, remapping=remapping)
-
-    return stmt
+    return apply_sorting(stmt, sorting, remapping=remapping)
 
 
 __all__ = [
     "apply_filters",
-    "apply_sorting",
     "apply_filters_and_sorting",
+    "apply_sorting",
 ]
