@@ -1,20 +1,31 @@
 from enum import Enum
-from typing import Dict, Any, Protocol, Tuple, Type, Callable, Union, Literal, List, Optional
+from typing import (
+    Any,
+    Callable,
+    Literal,
+    Optional,
+    Protocol,
+    Union,
+)
+
 from typing_extensions import TypeAlias
 
 from .fields import FilterField
 
 AbstractFilterOperator: TypeAlias = Enum
-FilterAliasGenerator: TypeAlias = Callable[[str, AbstractFilterOperator, Optional[str]], str]
+FilterAliasGenerator: TypeAlias = Callable[
+    [str, AbstractFilterOperator, Optional[str]],
+    str,
+]
 FilterPlace: TypeAlias = Callable[..., Any]
-FilterFieldDef: TypeAlias = Union[Type[Any], FilterField[Any]]
-FilterValues: TypeAlias = Dict[str, Dict[AbstractFilterOperator, Any]]
+FilterFieldDef: TypeAlias = Union[type[Any], FilterField[Any]]
+FilterValues: TypeAlias = dict[str, dict[AbstractFilterOperator, Any]]
 
 
 class FiltersResolver(Protocol):
-    __model__: Type[Any]
-    __defs__: Dict[str, Tuple[str, AbstractFilterOperator]]
-    __filters__: Dict[str, FilterField[Any]]
+    __model__: type[Any]
+    __defs__: dict[str, tuple[str, AbstractFilterOperator]]
+    __filters__: dict[str, FilterField[Any]]
 
     async def __call__(self, _: Any, /) -> FilterValues:  # pragma: no cover
         pass
@@ -22,12 +33,12 @@ class FiltersResolver(Protocol):
 
 SortingDirection: TypeAlias = Literal["asc", "desc"]
 SortingNulls: TypeAlias = Optional[Literal["bigger", "smaller"]]
-SortingValues: TypeAlias = List[Tuple[str, SortingDirection, SortingNulls]]
+SortingValues: TypeAlias = list[tuple[str, SortingDirection, SortingNulls]]
 
 
 class SortingResolver(Protocol):
     __tp__: Any
-    __defs__: Dict[str, Tuple[str, SortingDirection, SortingNulls]]
+    __defs__: dict[str, tuple[str, SortingDirection, SortingNulls]]
 
     async def __call__(self, _: Any, /) -> SortingValues:  # pragma: no cover
         pass
@@ -36,12 +47,12 @@ class SortingResolver(Protocol):
 __all__ = [
     "AbstractFilterOperator",
     "FilterAliasGenerator",
-    "FilterPlace",
     "FilterFieldDef",
+    "FilterPlace",
     "FilterValues",
     "FiltersResolver",
     "SortingDirection",
     "SortingNulls",
-    "SortingValues",
     "SortingResolver",
+    "SortingValues",
 ]
